xchain: more tests for historic nodes
